--- src/main/java/com/ict/jeju/wyy/dao/UserVO4.java
@@ -1,145 +1,0 @@
-package com.ict.jeju.wyy.dao;
-
-public class UserVO4 {
-	private String u_idx, u_name, u_id, u_pwd, u_birth, u_email, u_phone;
-	private String u_postcode, u_detail_addr, u_detail_addr2, u_addr, u_gender, u_state, u_regdate, u_profile_img;
-	private String c_idx, c_status;
-	private String like_idx, like_active;
-	private String bo_idx;
-	private String like_active_count, review_count;
-
-	public String getU_profile_img() {
-		return u_profile_img;
-	}
-	public void setU_profile_img(String u_profile_img) {
-		this.u_profile_img = u_profile_img;
-	}
-	public String getReview_count() {
-		return review_count;
-	}
-	public void setReview_count(String review_count) {
-		this.review_count = review_count;
-	}
-	public String getLike_active_count() {
-		return like_active_count;
-	}
-	public void setLike_active_count(String like_active_count) {
-		this.like_active_count = like_active_count;
-	}
-	public String getU_id() {
-		return u_id;
-	}
-	public void setU_id(String u_id) {
-		this.u_id = u_id;
-	}
-	public String getU_pwd() {
-		return u_pwd;
-	}
-	public void setU_pwd(String u_pwd) {
-		this.u_pwd = u_pwd;
-	}
-	public String getU_birth() {
-		return u_birth;
-	}
-	public void setU_birth(String u_birth) {
-		this.u_birth = u_birth;
-	}
-	public String getU_email() {
-		return u_email;
-	}
-	public void setU_email(String u_email) {
-		this.u_email = u_email;
-	}
-	public String getU_phone() {
-		return u_phone;
-	}
-	public void setU_phone(String u_phone) {
-		this.u_phone = u_phone;
-	}
-	public String getU_postcode() {
-		return u_postcode;
-	}
-	public void setU_postcode(String u_postcode) {
-		this.u_postcode = u_postcode;
-	}
-	public String getU_detail_addr() {
-		return u_detail_addr;
-	}
-	public void setU_detail_addr(String u_detail_addr) {
-		this.u_detail_addr = u_detail_addr;
-	}
-	public String getU_detail_addr2() {
-		return u_detail_addr2;
-	}
-	public void setU_detail_addr2(String u_detail_addr2) {
-		this.u_detail_addr2 = u_detail_addr2;
-	}
-	public String getU_addr() {
-		return u_addr;
-	}
-	public void setU_addr(String u_addr) {
-		this.u_addr = u_addr;
-	}
-	public String getU_gender() {
-		return u_gender;
-	}
-	public void setU_gender(String u_gender) {
-		this.u_gender = u_gender;
-	}
-	public String getU_state() {
-		return u_state;
-	}
-	public void setU_state(String u_state) {
-		this.u_state = u_state;
-	}
-	public String getU_regdate() {
-		return u_regdate;
-	}
-	public void setU_regdate(String u_regdate) {
-		this.u_regdate = u_regdate;
-	}
-	public String getU_idx() {
-		return u_idx;
-	}
-	public void setU_idx(String u_idx) {
-		this.u_idx = u_idx;
-	}
-	public String getU_name() {
-		return u_name;
-	}
-	public void setU_name(String u_name) {
-		this.u_name = u_name;
-	}
-	public String getC_idx() {
-		return c_idx;
-	}
-	public void setC_idx(String c_idx) {
-		this.c_idx = c_idx;
-	}
-	public String getC_status() {
-		return c_status;
-	}
-	public void setC_status(String c_status) {
-		this.c_status = c_status;
-	}
-	public String getLike_idx() {
-		return like_idx;
-	}
-	public void setLike_idx(String like_idx) {
-		this.like_idx = like_idx;
-	}
-	public String getLike_active() {
-		return like_active;
-	}
-	public void setLike_active(String like_active) {
-		this.like_active = like_active;
-	}
-	public String getBo_idx() {
-		return bo_idx;
-	}
-	public void setBo_idx(String bo_idx) {
-		this.bo_idx = bo_idx;
-	}
-	
-	
-}
