+++ src/main/resources/props/jdbc.properties
@@ -1,17 +1,17 @@
 # aws
-#jdbc.driver=org.mariadb.jdbc.Driver
-#jdbc.url=jdbc:mariadb://13.125.238.101:3306/team02_db
-#jdbc.user=team02
-#jdbc.pwd=1111
-#jdbc.maxActive=30
-#jdbc.maxIdle=20
-#jdbc.minIdle=10
-
-#oci
 jdbc.driver=org.mariadb.jdbc.Driver
-jdbc.url=jdbc:mariadb://127.0.0.1:3307/team02_db
+jdbc.url=jdbc:mariadb://13.125.238.101:3306/team02_db
 jdbc.user=team02
 jdbc.pwd=1111
 jdbc.maxActive=30
 jdbc.maxIdle=20
 jdbc.minIdle=10
+
+#oci
+#jdbc.driver=org.mariadb.jdbc.Driver
+#jdbc.url=jdbc:mariadb://127.0.0.1:3307/team02_db
+#jdbc.user=team02
+#jdbc.pwd=1111
+#jdbc.maxActive=30
+#jdbc.maxIdle=20
+#jdbc.minIdle=10
