+++ src/main/webapp/WEB-INF/views/lsh_view/admin_join_page.jsp
@@ -171,7 +171,7 @@
 				<td><span class="material-symbols-outlined">manage_accounts</span><br><a href="user_list.do">회원관리</a></td>
 				<td><span class="material-symbols-outlined">mark_chat_read</span><br><a href="admin_list2.do">답변</a></td>
 				<td><span class="material-symbols-outlined">mark_chat_unread</span><br><a href="admin_list.do">미답변</a></td>
-				<td><span class="material-symbols-outlined">calendar_add_on</span><br><a href="admin_insert">일정 추가</a></td>
+				<td><span class="material-symbols-outlined">calendar_add_on</span><br><a href="admin_insert">장소 추가</a></td>
 				<c:if test="${adminVO.a_status == '1'}">
 					<td><span class="material-symbols-outlined">person_add</span><br><a href="admin_join.do">관리자 생성</a></td>
 				</c:if>
@@ -187,7 +187,7 @@
 				<li><span class="material-symbols-outlined">manage_accounts</span><a href="user_list.do">회원관리</a></li>
 				<li><span class="material-symbols-outlined">mark_chat_read</span><a href="admin_list2.do">답변</a></li>
 				<li><span class="material-symbols-outlined">mark_chat_unread</span><a href="admin_list.do">미답변</a></li>
-				<li><span class="material-symbols-outlined">calendar_add_on</span><a href="admin_insert">일정 추가</a></li>
+				<li><span class="material-symbols-outlined">calendar_add_on</span><a href="admin_insert">장소 추가</a></li>
 				<c:if test="${adminVO.a_status == '1'}">
 					<li><span class="material-symbols-outlined">person_add</span><a href="admin_join.do">관리자 생성</a></li>
 				</c:if>
